+++ src/app/fyle/add-edit-expense/add-edit-expense.page.ts
@@ -3021,16 +3021,7 @@ export class AddEditExpensePage implements OnInit {
             that.editExpense('SAVE_EXPENSE').subscribe(() => this.goBack());
           }
         } else {
-          that.fg.markAllAsTouched();
-          const formContainer = that.formContainer.nativeElement as HTMLElement;
-          if (formContainer) {
-            const invalidElement = formContainer.querySelector('.ng-invalid');
-            if (invalidElement) {
-              invalidElement.scrollIntoView({
-                behavior: 'smooth',
-              });
-            }
-          }
+          that.showFormValidationErrors();
 
           if (invalidPaymentMode) {
             that.invalidPaymentMode = true;
@@ -3060,16 +3051,7 @@ export class AddEditExpensePage implements OnInit {
             });
           }
         } else {
-          that.fg.markAllAsTouched();
-          const formContainer = that.formContainer.nativeElement as HTMLElement;
-          if (formContainer) {
-            const invalidElement = formContainer.querySelector('.ng-invalid');
-            if (invalidElement) {
-              invalidElement.scrollIntoView({
-                behavior: 'smooth',
-              });
-            }
-          }
+          that.showFormValidationErrors();
           if (invalidPaymentMode) {
             that.invalidPaymentMode = true;
             setTimeout(() => {
@@ -3102,16 +3084,7 @@ export class AddEditExpensePage implements OnInit {
         });
       }
     } else {
-      that.fg.markAllAsTouched();
-      const formContainer = that.formContainer.nativeElement as HTMLElement;
-      if (formContainer) {
-        const invalidElement = formContainer.querySelector('.ng-invalid');
-        if (invalidElement) {
-          invalidElement.scrollIntoView({
-            behavior: 'smooth',
-          });
-        }
-      }
+      that.showFormValidationErrors();
     }
   }
 
@@ -3137,16 +3110,7 @@ export class AddEditExpensePage implements OnInit {
         });
       }
     } else {
-      that.fg.markAllAsTouched();
-      const formContainer = that.formContainer.nativeElement as HTMLElement;
-      if (formContainer) {
-        const invalidElement = formContainer.querySelector('.ng-invalid');
-        if (invalidElement) {
-          invalidElement.scrollIntoView({
-            behavior: 'smooth',
-          });
-        }
-      }
+      that.showFormValidationErrors();
     }
   }
 
@@ -4006,8 +3970,31 @@ export class AddEditExpensePage implements OnInit {
       });
   }
 
+  getDeleteReportParams(
+    config: { header: string; body: string; ctaText: string; ctaLoadingText: string },
+    removeExpenseFromReport: boolean,
+    reportId?: string
+  ) {
+    return {
+      component: FyDeleteDialogComponent,
+      cssClass: 'delete-dialog',
+      backdropDismiss: false,
+      componentProps: {
+        header: config.header,
+        body: config.body,
+        ctaText: config.ctaText,
+        ctaLoadingText: config.ctaLoadingText,
+        deleteMethod: () => {
+          if (removeExpenseFromReport) {
+            return this.reportService.removeTransaction(reportId, this.activatedRoute.snapshot.params.id);
+          }
+          return this.transactionService.delete(this.activatedRoute.snapshot.params.id);
+        },
+      },
+    };
+  }
+
   async deleteExpense(reportId?: string) {
-    const id = this.activatedRoute.snapshot.params.id;
     const removeExpenseFromReport = reportId && this.isRedirectedFromReport;
     const header = removeExpenseFromReport ? 'Remove Expense' : 'Delete Expense';
     const body = removeExpenseFromReport
@@ -4016,23 +4003,18 @@ export class AddEditExpensePage implements OnInit {
     const ctaText = removeExpenseFromReport ? 'Remove' : 'Delete';
     const ctaLoadingText = removeExpenseFromReport ? 'Removing' : 'Deleting';
 
-    const deletePopover = await this.popoverController.create({
-      component: FyDeleteDialogComponent,
-      cssClass: 'delete-dialog',
-      backdropDismiss: false,
-      componentProps: {
-        header,
-        body,
-        ctaText,
-        ctaLoadingText,
-        deleteMethod: () => {
-          if (removeExpenseFromReport) {
-            return this.reportService.removeTransaction(reportId, id);
-          }
-          return this.transactionService.delete(id);
+    const deletePopover = await this.popoverController.create(
+      this.getDeleteReportParams(
+        {
+          header,
+          body,
+          ctaText,
+          ctaLoadingText,
         },
-      },
-    });
+        removeExpenseFromReport,
+        reportId
+      )
+    );
 
     await deletePopover.present();
     const { data } = await deletePopover.onDidDismiss();
@@ -4332,7 +4314,7 @@ export class AddEditExpensePage implements OnInit {
     this.selectedProject$.complete();
   }
 
-  private async showSizeLimitExceededPopover() {
+  async showSizeLimitExceededPopover() {
     const sizeLimitExceededPopover = await this.popoverController.create({
       component: PopupAlertComponent,
       componentProps: {
